refactor(tests): rename api tests
