+++ __tests__/integration/api-mark-change-data.spec.ts
@@ -1,19 +1,19 @@
-import { basicIntervalChartChangeSize as render } from '../plots/api/basic-interval-chart-change-size';
+import { markChangeData as render } from '../plots/api/mark-change-data';
 import { createNodeGCanvas } from './utils/createNodeGCanvas';
 import { sleep } from './utils/sleep';
 import './utils/useSnapshotMatchers';
 
-describe('mark.changeSize', () => {
+describe('mark.changeData', () => {
   const canvas = createNodeGCanvas(640, 480);
 
-  it('mark.changeSize(width, height) should rerender expected chart', async () => {
+  it('mark.changeData(width, height) should rerender expected chart', async () => {
     const { finished, button, chart } = render({
       canvas,
       container: document.createElement('div'),
     });
     await finished;
     button.dispatchEvent(new CustomEvent('click'));
-    await new Promise<void>((resolve) => chart.on('afterchangesize', resolve));
+    await new Promise<void>((resolve) => chart.on('afterrender', resolve));
     const dir = `${__dirname}/snapshots/api`;
     await sleep(20);
     await expect(canvas).toMatchCanvasSnapshot(dir, render.name);
