+++ __tests__/integration/snapshots/api/chart-auto-fit/step0.png
