+++ __tests__/integration/snapshots/api/chart-auto-fit/step1.png
