+++ __tests__/integration/snapshots/api/chartChangeSize.png
