+++ __tests__/integration/snapshots/api/chartRender.png
