+++ __tests__/integration/snapshots/api/registerDataTransform.png
