+++ __tests__/integration/snapshots/api/registerShape.png
