+++ __tests__/plots/api/index.ts
@@ -1,5 +1,5 @@
-export { basicIntervalRegisterDataTransform } from './basic-interval-register-data-transform';
-export { basicIntervalRegisterShape } from './basic-interval-register-shape';
-export { basicIntervalChartChangeSize } from './basic-interval-chart-change-size';
-export { basicIntervalChartAutoFit } from './basic-interval-chart-auto-fit';
-export { basicIntervalMarkChangeData } from './basic-interval-mark-change-data';
+export { registerDataTransform } from './register-data-transform';
+export { registerShape } from './register-shape';
+export { chartChangeSize } from './chart-change-size';
+export { chartAutoFit } from './chart-auto-fit';
+export { markChangeData } from './mark-change-data';
